Minor tweak to remove warnings
git-svn-id: https://swig.svn.sourceforge.net/svnroot/swig/trunk/SWIG@4773 626c5289-ae23-0410-ae9c-e8d60b6d4f22

+++ Source/Modules/s-exp.cxx
@@ -187,7 +187,7 @@ public:
     // Following is a silly hack.  It works around the limitation of
     // DOH's hash tables that only work with string keys!
     char address[16];
-    sprintf(address, "%x%c", (unsigned int)obj, list_p ? 'L' : 'O');
+    sprintf(address, "%x%c", (unsigned long)obj, list_p ? 'L' : 'O');
     DOH *placeholder = Getattr(print_circle_hash, address);
     if (placeholder) {
       Printv(out, placeholder, NIL);
